Changed USer from concrete class to interface, and added AdopterUser interface and its concrete class to follow clean arch rules like SRP and OCP. Still partial implementation since dependent classes are not yet implemented
